how_to_create_a_chain_resources/requirements.txt
python-dotenv
langchain
langchain-openai
langserve[all]